+++ jgclark.DailyJournal/src/index.js
@@ -1,45 +1,64 @@
 // @flow
 
 //-----------------------------------------------------------------------------
-// Daily Journal commands
+// Journalling commands
 // Jonathan Clark
-// Last updated 18.7.22 for v0.12.1
+// Last updated 21.8.22 for v0.13.0
 //-----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation
 import pluginJson from '../plugin.json'
-import { log, logError } from '@helpers/dev'
-import { updateSettingData } from '@helpers/NPConfiguration'
+import { JSP, logDebug, logInfo, logError } from "@helpers/dev"
+import { pluginUpdated, updateSettingData } from '@helpers/NPConfiguration'
 import { showMessage } from '@helpers/userInput'
 
-export { dayStart, dayReview, todayStart } from './journal'
+export {
+  dayStart,
+  dailyJournalQuestions,
+  monthlyJournalQuestions,
+  quarterlyJournalQuestions,
+  todayStart,
+  weeklyJournalQuestions
+} from './journal'
 
 export function init(): void {
-  // In the background, see if there is an update to the plugin to install, and if so let user know
-  DataStore.installOrUpdatePluginsByID([pluginJson['plugin.id']], false, false, false)
+  try {
+    // Check for the latest version of the plugin, and if a minor update is available, install it and show a message
+    DataStore.installOrUpdatePluginsByID([pluginJson['plugin.id']], false, false, false).then((r) =>
+      pluginUpdated(pluginJson, r),
+    )
+  } catch (error) {
+    logError(pluginJson, JSP(error))
+  }
 }
 
 export function onSettingsUpdated(): void {
   // Placeholder only to stop error in logs
 }
 
-const configKey = 'dailyJournal'
+const pluginID = 'jgclark.DailyJournal'
 
-export async function onUpdateOrInstall(): Promise<void> {
+// test the update mechanism, including display to user
+export function testUpdate(): void {
+  onUpdateOrInstall(true) // force update mechanism to fire
+}
+
+export async function onUpdateOrInstall(testUpdate: boolean = false): Promise<void> {
   try {
-    log(pluginJson, `${configKey}: onUpdateOrInstall running`)
-    // Try updating settings data
-    const updateSettings = updateSettingData(pluginJson)
-    log(pluginJson, `${configKey}: onUpdateOrInstall updateSettingData code: ${updateSettings}`)
+    logInfo(pluginID, `onUpdateOrInstall ...`)
+    let updateSettingsResult = updateSettingData(pluginJson)
+    logInfo(pluginID, `- updateSettingData code: ${updateSettingsResult}`)
+
+    if (testUpdate) {
+      updateSettingsResult = 1 // updated
+      logDebug(pluginID, '- forcing pluginUpdated() to run ...')
+    }
 
     // Tell user the plugin has been updated
-    if (pluginJson['plugin.lastUpdateInfo'] !== 'undefined') {
-      await showMessage(pluginJson['plugin.lastUpdateInfo'], 'OK, thanks',
-        `Plugin ${pluginJson['plugin.name']}\nupdated to v${pluginJson['plugin.version']}`
-      )
-    }
+    await pluginUpdated(pluginJson, { code: updateSettingsResult, message: 'unused?' })
+
   } catch (error) {
-    logError(pluginJson, error)
+    logError(pluginID, error.message)
   }
-  log(pluginJson, `${configKey}: onUpdateOrInstall finished`)
+  logInfo(pluginID, `- finished`)
 }